requirements.txt
aiodns
aiohttp
click
jinja2
jsonschema
pyyaml
